requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2025.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0 #Corrected entry
plotly==5.24.1
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
streamlit==1.34.0
streamlit-lottie==0.0.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
youtube-dl==2021.12.17